refactor: move PriceConverter class to a separate file and streamline message handling
